+++ sdk/python/pulumi_openstack/loadbalancer/outputs.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = [
@@ -17,22 +17,22 @@ __all__ = [
 class MembersMember(dict):
     def __init__(__self__, *,
                  address: str,
-                 protocol_port: float,
+                 protocol_port: int,
                  admin_state_up: Optional[bool] = None,
                  id: Optional[str] = None,
                  name: Optional[str] = None,
                  subnet_id: Optional[str] = None,
-                 weight: Optional[float] = None):
+                 weight: Optional[int] = None):
         """
         :param str address: The IP address of the members to receive traffic from
                the load balancer.
-        :param float protocol_port: The port on which to listen for client traffic.
+        :param int protocol_port: The port on which to listen for client traffic.
         :param bool admin_state_up: The administrative state of the member.
                A valid value is true (UP) or false (DOWN). Defaults to true.
         :param str id: The unique ID for the members.
         :param str name: Human-readable name for the member.
         :param str subnet_id: The subnet in which to access the member.
-        :param float weight: A positive integer value that indicates the relative
+        :param int weight: A positive integer value that indicates the relative
                portion of traffic that this members should receive from the pool. For
                example, a member with a weight of 10 receives five times as much traffic
                as a member with a weight of 2. Defaults to 1.
@@ -61,7 +61,7 @@ class MembersMember(dict):
 
     @property
     @pulumi.getter(name="protocolPort")
-    def protocol_port(self) -> float:
+    def protocol_port(self) -> int:
         """
         The port on which to listen for client traffic.
         """
@@ -102,7 +102,7 @@ class MembersMember(dict):
 
     @property
     @pulumi.getter
-    def weight(self) -> Optional[float]:
+    def weight(self) -> Optional[int]:
         """
         A positive integer value that indicates the relative
         portion of traffic that this members should receive from the pool. For
